requirements.txt
snowflake-snowpark-python==1.4.0
streamlit==1.22.0
pandas==2.0.1
openpyxl==3.1.2